requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# bot.py: 3
# modules\botlist.py: 2
# modules\developer.py: 5
# modules\error_handler.py: 2
# modules\image.py: 2
# modules\informational.py: 2
# modules\utils\hastebin.py: 1
aiohttp == 3.3.2

# bot.py: 2,8
# modules\_help.py: 1,3
# modules\botlist.py: 3
# modules\developer.py: 1,10
# modules\error_handler.py: 1,5
# modules\image.py: 1,4
# modules\informational.py: 1,6
# modules\systeminfo.py: 8
# modules\utils\checks.py: 1
# modules\utils\paginator.py: 6
discord.py == 1.0.0a1483+gec3435b

# modules\informational.py: 4
# modules\systeminfo.py: 1
psutil == 5.4.6

# bot.py: 11
pyfiglet == 0.7.5

# bot.py: 6
# modules\developer.py: 8
rethinkdb == 2.3.0.post6
